update web bin
